requirements.txt
pytesseract
Pillow
pdf2image
pandas
camelot-py